Outsource FieldTrait
